bot/requirements.txt
discord.py==2.3.2
anthropic>=0.39.0
letta-client>=0.1.0
apscheduler==3.10.4
python-dotenv==1.0.0
pytz>=2024.1

# Claude Agent SDK (uses Max subscription via OAuth)
claude-agent-sdk>=0.1.0
anyio>=4.0.0
